application/requirements.txt
django
djangorestframework
pandas
numpy
django-cors-headers
